Move helper functions in oak_loader to a separate submodule
